thriftmux depends on thrift to know how to import thrift classes
at some point it may be worth deduplicating some more stuff between
thriftmux & thrift. for the moment i think it's ok as it makes it a
little easier to use thriftmux.

it might make sense conceptually for a thriftmux server to `serve`
a thrift `service` but that makes the API a little more complicated?